+++ noodles-cram/examples/cram_view.rs
@@ -30,7 +30,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut reader = File::open(src).map(cram::Reader::new)?;
     reader.read_file_definition()?;
 
-    let header: sam::Header = reader.read_file_header()?.parse()?;
+    let header = reader.read_file_header()?;
 
     let stdout = io::stdout().lock();
     let mut writer = sam::Writer::new(BufWriter::new(stdout));